kedro_mlops_mlflow/{{ cookiecutter.repo_name }}/pyproject.toml
[tool.kedro]
package_name = "{{ cookiecutter.python_package }}"
project_name = "{{ cookiecutter.project_name }}"
project_version = "{{ cookiecutter.kedro_version }}"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
known_third_party = "kedro"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/{{ cookiecutter.python_package }} -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.poetry]
name = "{{ cookiecutter.project_name }}"
version = "0.1.0"
description = ""
authors = ["SeraphnetAI"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
isort = "5.0"
kedro = "0.18.1"
kedro-telemetry = "0.2.0"
kedro-docker = "^0.3.0"
db-dtypes = "^1.0.2"
pytest-cov = "3.0"
pytest-mock = "2.0"
Jinja2 = "3.1.0"
pandas = "^1.4.3"
numpy = ">=1.21.0,<1.23"
requests = "2.25.1"
pyarrow = "^6.0.1"
oauth2client = "^4.1.3"
google-cloud = "^0.34.0"
google-auth = ">=1.6.1,<2"
google-api-python-client = ">=1.7.8,<2"
google-cloud-bigquery = "^3.2.0"
google-cloud-storage = ">=1.20.0,<2"
google-cloud-pubsub = "2.13.6"
appengine-python-standard = "^1.0.0"
google-cloud-artifact-registry = "^1.3.3"
google-cloud-build = "^3.9.3"
google-cloud-functions = "^1.8.3"
google-cloud-run = "^0.4.2"
google-cloud-scheduler = "^2.7.3"
cloud-sql-python-connector = "^0.9.0"
google-cloud-compute = "^1.6.1"
google-cloud-dataflow-client = "^0.6.2"
google-cloud-dataproc = "^5.0.3"
gcsfs = "2022.1.0"
jupyter = "^1.0.0"
jupyterlab = "^3.4.6"
nbstripout = "^0.6.0"
jupytext = "^1.14.1"
kedro-mlflow = "^0.11.3" 

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
